+++ test/smoke/src/areas/positron/apps/shiny.test.ts
@@ -4,9 +4,10 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { Application, PositronPythonFixtures, PositronRFixtures } from '../../../../../automation';
-import { setupAndStartApp } from '../../../positronUtils';
+import { setupAndStartApp } from '../../../test-hooks';
 import { expect } from '@playwright/test';
 import { join } from 'path';
+
 
 
 describe('Shiny Application', () => {
