+++ test/smoke/src/areas/positron/dataexplorer/data-explorer-100x100.test.ts
@@ -6,7 +6,7 @@
 import * as fs from 'fs';
 import { expect } from '@playwright/test';
 import { Application, PositronPythonFixtures, PositronRFixtures } from '../../../../../automation';
-import { setupAndStartApp } from '../../../positronUtils';
+import { setupAndStartApp } from '../../../test-hooks';
 import { join } from 'path';
 
 describe('Data Explorer 100x100', function () {
