+++ test/smoke/src/areas/positron/dataexplorer/sparklinesTrend.test.ts
@@ -5,7 +5,7 @@
 
 import { expect } from '@playwright/test';
 import { Application, PositronPythonFixtures, PositronRFixtures } from '../../../../../automation';
-import { setupAndStartApp } from '../../../positronUtils';
+import { setupAndStartApp } from '../../../test-hooks';
 
 describe('Data Explorer #web', () => {
 	setupAndStartApp();
