+++ test/smoke/src/areas/positron/r-pkg-development/r-pkg-development.test.ts
@@ -6,7 +6,7 @@
 import path = require('path');
 import { Application, PositronRFixtures, PositronUserSettingsFixtures } from '../../../../../automation';
 import { expect } from '@playwright/test';
-import { setupAndStartApp } from '../../../positronUtils';
+import { setupAndStartApp } from '../../../test-hooks';
 
 let logger;
