+++ test/smoke/src/areas/positron/rmarkdown/rmarkdown.test.ts
@@ -6,7 +6,7 @@
 import { Application, PositronRFixtures } from '../../../../../automation';
 import { join } from 'path';
 import { expect } from '@playwright/test';
-import { setupAndStartApp } from '../../../positronUtils';
+import { setupAndStartApp } from '../../../test-hooks';
 
 
 describe('RMarkdown #web', () => {
