+++ test/smoke/src/areas/positron/welcome/welcome.test.ts
@@ -6,7 +6,7 @@
 
 import { expect } from '@playwright/test';
 import { Application, PositronPythonFixtures, PositronRFixtures } from '../../../../../automation';
-import { setupAndStartApp } from '../../../positronUtils';
+import { setupAndStartApp } from '../../../test-hooks';
 
 describe('Welcome Page', () => {
 	setupAndStartApp();
