+++ test/smoke/src/test-hooks.ts
@@ -5,7 +5,7 @@
 
 import { join } from 'path';
 import * as path from 'path';
-import { Logger, } from '../../automation';
+import { Logger, } from '../../automation/out';
 import { installAllHandlers, } from './utils';
 import { createLogger } from './test-runner/logger';
 import { OPTS } from './test-runner/config';
